Add IssueRequest model to handle issue data in API
